requirements.txt
jinja2>=2.7
pyyaml>=5.1.2
gitpython
